Cambiamenti: modificato nav_home in nav_home_studente.xml e nav_home_professore.xml modificato fragment_home.xml in fragment_home_studente.xml e fragment_home_professore.xml in modo tale da distinguere le due viste create le classi HomeProfessoreFragment e HomeStudenteFragment creata la bottom_nav_menu.xml contenente i pulsanti home e profilo, i questo modo può essere utilizzata in tutti i fragment post-login immediatamente.
