+++ app/src/main/res/drawable/outline_assignment_ind_24.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#004070"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M19,3h-4.18C14.4,1.84 13.3,1 12,1s-2.4,0.84 -2.82,2L5,3c-1.1,0 -2,0.9 -2,2v14c0,1.1 0.9,2 2,2h14c1.1,0 2,-0.9 2,-2L21,5c0,-1.1 -0.9,-2 -2,-2zM12,2.75c0.22,0 0.41,0.1 0.55,0.25 0.12,0.13 0.2,0.31 0.2,0.5 0,0.41 -0.34,0.75 -0.75,0.75s-0.75,-0.34 -0.75,-0.75c0,-0.19 0.08,-0.37 0.2,-0.5 0.14,-0.15 0.33,-0.25 0.55,-0.25zM19,19L5,19L5,5h14v14zM12,6c-1.65,0 -3,1.35 -3,3s1.35,3 3,3 3,-1.35 3,-3 -1.35,-3 -3,-3zM12,10c-0.55,0 -1,-0.45 -1,-1s0.45,-1 1,-1 1,0.45 1,1 -0.45,1 -1,1zM6,16.47L6,18h12v-1.53c0,-2.5 -3.97,-3.58 -6,-3.58s-6,1.07 -6,3.58zM8.31,16c0.69,-0.56 2.38,-1.12 3.69,-1.12s3.01,0.56 3.69,1.12L8.31,16z"/>
+</vector>
