+++ app/src/main/res/drawable/outline_summarize_24.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#004070"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M15,3H5C3.9,3 3.01,3.9 3.01,5L3,19c0,1.1 0.89,2 1.99,2H19c1.1,0 2,-0.9 2,-2V9L15,3zM5,19V5h9v5h5v9H5zM9,8c0,0.55 -0.45,1 -1,1S7,8.55 7,8s0.45,-1 1,-1S9,7.45 9,8zM9,12c0,0.55 -0.45,1 -1,1s-1,-0.45 -1,-1s0.45,-1 1,-1S9,11.45 9,12zM9,16c0,0.55 -0.45,1 -1,1s-1,-0.45 -1,-1s0.45,-1 1,-1S9,15.45 9,16z"/>
+</vector>
