+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:title="@string/home"
+        android:id="@+id/home"
+        android:icon="@drawable/outline_home_24"
+        />
+    <item android:title="@string/profilo"
+        android:id="@+id/profilo"
+        android:icon="@drawable/baseline_person_outline_24"
+        />
+
+</menu>
