+++ app/src/main/res/navigation/nav_home_studente.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_home_studente"
+    app:startDestination="@id/home_fragment">
+
+    <fragment
+        android:id="@+id/home_fragment"
+        android:name="com.laureapp.ui.home.HomeStudenteFragment"
+        android:label="Laureapp"
+        tools:layout="@layout/fragment_home_studente" >
+        <action
+            android:id="@+id/action_home_fragment_to_registerStudenteFragment"
+           />
+
+    </fragment>
+
+
+</navigation>
