+++ app/src/main/res/values/strings.xml
@@ -28,4 +28,19 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="matricola">Matricola</string>
+    <string name="home">Home</string>
+    <string name="tesi">Tesi</string>
+    <string name="messaggi">Messaggi</string>
+    <string name="classifica">Classifica</string>
+    <string name="social">Social</string>
+    <string name="tesi_description">tesi_description</string>
+    <string name="chat_description">chat</string>
+    <string name="classifica_description" />
+    <string name="social_description" />
+    <string name="messaggi_description" />
+    <string name="task">Task</string>
+    <string name="task_description" />
+    <string name="tesisti" />
+    <string name="tesisti_description" />
+    <string name="tesisti2">Tesisti</string>
 </resources>
